codeInOne_API/quickstart/serializers.py
from django.contrib.auth.models import User, Group
from rest_framework import serializers
from codeInOne_API.quickstart.models import Challenge


class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ['url', 'username', 'email', 'groups']


class GroupSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Group
        fields = ['url', 'name']

class ChallengeSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Challenge
        fields = ['challengeName', 'challengeDesc', 'release_date']